+++ cognite/powerops/preprocessor/main.py
@@ -1,4 +1,3 @@
-import os
 from typing import Optional
 from uuid import uuid4
 
@@ -7,12 +6,11 @@ from cognite.client import CogniteClient
 
 from cognite.powerops.preprocessor import knockoff_logging as logging
 from cognite.powerops.preprocessor.cogreader import CogReader
-from cognite.powerops.preprocessor.utils import initialize_cognite_client
 
 logger = logging.getLogger(__name__)
 
 
-def _main(
+def main(
     client: CogniteClient,
     fdm_space_external_id: str,
     fdm_case_external_id: str,
@@ -31,7 +29,7 @@ def _main(
 
     extra_files_md = cog_reader.get_extra_files_metadata()
     mapping_files_md = cog_reader.get_mapping_files_metadata()
-    cut_file_md = cog_reader.get_cut_file_metadata()
+    cut_files_md = cog_reader.get_cut_files_metadata()
 
     xid = f"cog_shop_preprocessor/{fdm_space_external_id}/{fdm_case_external_id}/{str(uuid4())}"
 
@@ -50,36 +48,7 @@ def _main(
 
     return {
         "cog_shop_case_file": cog_reader.file_metadata_to_dict(cog_shop_case_file_md),
-        "cut_file": cut_file_md,
+        "cut_files": cut_files_md,
         "mapping_files": mapping_files_md,
         "extra_files": extra_files_md,
     }
-
-
-def main(
-    output_data_set_id: int,
-    client: CogniteClient = None,
-    fdm_space_external_id: Optional[str] = None,
-    fdm_case_external_id: Optional[str] = None,
-    fdm_model_version: Optional[int] = None,
-) -> dict:
-    client = client if client is not None else initialize_cognite_client()
-
-    fdm_space_external_id = (
-        fdm_space_external_id if fdm_space_external_id is not None else os.getenv("FDM_SPACE_EXTERNAL_ID")
-    )
-    fdm_case_external_id = (
-        fdm_case_external_id if fdm_case_external_id is not None else os.getenv("FDM_CASE_EXTERNAL_ID")
-    )
-
-    result = _main(
-        client=client,
-        fdm_space_external_id=fdm_space_external_id,  # type: ignore[arg-type]
-        fdm_case_external_id=fdm_case_external_id,  # type: ignore[arg-type]
-        output_data_set_id=output_data_set_id,
-        fdm_model_version=fdm_model_version,
-    )
-
-    logger.info(f"main() returned: {result!r}")
-
-    return result